Fixed template, add missing options to manifests
The new configuration options where missed being added to the puppet
files.  Correctly fix this to ensure we can use the module within 1.0
and later.
Fixed the template from having quotes around a int which blew up
influxdb.
Stripped down the blank lines on the config file to be a little bit more
neater